:+1: Add support for scrolling weightView #12
